Fix footer to the bottom
In cases where the content was short enough to not take up
the full height of the window, the footer would float off
the bottom because the body element would be shorter than
the screen. `height: 100vh;` sets the height of the body
to be the height of the viewport if it is shorter.