Create central column structure
In the reference image, many elements are organized into a
central "column" that spans the full height of the sight. For example,
the Header Logo and the CTA box both share a left edge.

This is a little complicated to do because each section is a div
that spans the entire page horizontally, so I can't make a single
"column" div to define this left (and right) edge because it would
be intersecting the horizontal section divs.

I the column layout by creating separate .column divs nested inside
each section. The .column divs are each centered and share a width
via a shared `flex-basis`.

In general, the CSS is now organized such that each section centers
its fixed-basis .column div. Then each section's .column div defines
it's own internal structure to achieve the desired layout for that
section.